Add Python agent support
This adds a new library in this repository being the
liblttng-ust-python-agent. It contains the Python agent and a library
for the tracepoints used by the agent.

To enable the python agent, use --enable-python-agent to build it.

This new agent is bound to the LTTng python domain in lttng-tools which
implements the -p,--python option for agent's command.

It works and tested for python 2.7.x and 3.x

Signed-off-by: David Goulet <dgoulet@efficios.com>